src/data/dataset_preparation/requirements.txt
opencv-python==4.8.0.76
Pillow==8.3.2
scikit-learn==1.3.0
tensorflow==2.11.0
pandas==1.4.4
tqdm==4.63.0
py7zr==0.20.0
imagecorruptions
imgaug
numpy<2
py7zr